+++ inc/models/GridModel.h
@@ -15,9 +15,9 @@ class GridModel : public QObject
     Q_OBJECT
 
     Q_PROPERTY(QVector<QObject *> grid READ grid NOTIFY gridChanged)
-    Q_PROPERTY(unsigned int rows READ rows NOTIFY rowsChanged)
-    Q_PROPERTY(unsigned int columns READ columns NOTIFY columnsChanged)
-    Q_PROPERTY(unsigned int mineCount READ mineCount NOTIFY mineCountChanged)
+    Q_PROPERTY(quint64 rows READ rows NOTIFY rowsChanged)
+    Q_PROPERTY(quint64 columns READ columns NOTIFY columnsChanged)
+    Q_PROPERTY(quint64 mineCount READ mineCount NOTIFY mineCountChanged)
 
     public:
         //delete default constructor
@@ -25,24 +25,24 @@ class GridModel : public QObject
 
         // custom constructor
         GridModel(const QVector<CellModel *> grid,
-                  const unsigned int rows,
-                  const unsigned int columns,
-                  const unsigned int mineCount);
+                  const quint64 rows,
+                  const quint64 columns,
+                  const quint64 mineCount);
 
         // default destructor
         ~GridModel() = default;
 
         // property getters
         QVector<QObject *> grid();
-        unsigned int rows();
-        unsigned int columns();
-        unsigned int mineCount();
+        quint64 rows();
+        quint64 columns();
+        quint64 mineCount();
 
         // property setters
         void setGrid(const QVector<CellModel *> grid);
-        void setRows(const unsigned int rows);
-        void setColumns(const unsigned int columns);
-        void setMineCount(const unsigned int mineCount);
+        void setRows(const quint64 rows);
+        void setColumns(const quint64 columns);
+        void setMineCount(const quint64 mineCount);
 
     signals:
         void gridChanged();
@@ -54,9 +54,9 @@ class GridModel : public QObject
 
         // properties
         QVector<CellModel *> mGrid;
-        unsigned int mRows;
-        unsigned int mColumns;
-        unsigned int mMineCount;
+        quint64 mRows;
+        quint64 mColumns;
+        quint64 mMineCount;
 
 };
